+++ datanode/sqlsubscribers/position.go
@@ -54,6 +54,11 @@ type settleDistressed interface {
 	Margin() *num.Uint
 }
 
+type ordersClosed interface {
+	MarketID() string
+	Parties() []string
+}
+
 type settleMarket interface {
 	positionEventBase
 	SettledPrice() *num.Uint
@@ -64,6 +69,7 @@ type PositionStore interface {
 	Add(context.Context, entities.Position) error
 	GetByMarket(ctx context.Context, marketID string) ([]entities.Position, error)
 	GetByMarketAndParty(ctx context.Context, marketID string, partyID string) (entities.Position, error)
+	GetByMarketAndParties(ctx context.Context, marketID string, parties []string) ([]entities.Position, error)
 	Flush(ctx context.Context) error
 }
 
@@ -92,6 +98,7 @@ func (p *Position) Types() []events.Type {
 		events.LossSocializationEvent,
 		events.SettleMarketEvent,
 		events.TradeEvent,
+		events.DistressedOrdersClosedEvent,
 	}
 }
 
@@ -112,9 +119,27 @@ func (p *Position) Push(ctx context.Context, evt events.Event) error {
 		return p.handleSettleMarket(ctx, event)
 	case tradeEvent:
 		return p.handleTradeEvent(ctx, event)
+	case ordersClosed:
+		return p.handleOrdersClosedEvent(ctx, event)
 	default:
 		return errors.Errorf("unknown event type %s", evt.Type().String())
 	}
+}
+
+func (p *Position) handleOrdersClosedEvent(ctx context.Context, event ordersClosed) error {
+	p.mutex.Lock()
+	defer p.mutex.Unlock()
+	// @TODO implement this
+	positions, err := p.store.GetByMarketAndParties(ctx, event.MarketID, event.Parties())
+	if err != nil {
+		return errors.Wrap(err, "error getting positions")
+	}
+	for _, pos := range positions {
+		pos.UpdateOrdersClosed()
+		// an update of an existing position here can't really fail
+		_ = p.updatePosition(ctx, pos)
+	}
+	return nil
 }
 
 func (p *Position) handleTradeEvent(ctx context.Context, event tradeEvent) error {